Databases fixes: context session manager + migration to PostgreSQL + async io support.

+++ app/models/conversation.py
@@ -1,14 +1,14 @@
 from models.base import Base
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, Integer, BIGINT
 from sqlalchemy.orm import relationship
 
 
 class Conversation(Base):
     __tablename__ = "conversations"
     id = Column(Integer, primary_key=True)
-    user1_id = Column(Integer, ForeignKey("users.id"))
+    user1_id = Column(BIGINT, ForeignKey("users.id"))
     user1_profile_version = Column(Integer)
-    user2_id = Column(Integer, ForeignKey("users.id"))
+    user2_id = Column(BIGINT, ForeignKey("users.id"))
     user2_profile_version = Column(Integer)
     start_time = Column(DateTime)
     end_time = Column(DateTime, nullable=True, default=None)
