+++ app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Integer, String, Boolean, BIGINT
 from models.base import Base
 from sqlalchemy.orm import relationship
 from models.message import Message
@@ -8,7 +8,7 @@ from models.profile_data import ProfileData
 class User(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BIGINT, primary_key=True)
     username = Column(String, unique=True)
     is_active = Column(Boolean, default=True)
     is_ready_to_chat = Column(Boolean, default=False)
